% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rude_answer.R
\name{bye}
\alias{bye}
\title{A Bye function}
\usage{
bye(good_day = TRUE)
}
\arguments{
\item{good_day}{Did you have a good day? Defaults to TRUE.}
}
\description{
This function allows you be sligthly rude or very rude when saying good-bye to someone.
}
\examples{
bye()
}
\keyword{rude}
