% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rude_answer.R
\name{hello}
\alias{hello}
\title{A Hello function}
\usage{
hello(good_day = TRUE)
}
\arguments{
\item{good_day}{Did you have a good day? Defaults to TRUE.}
}
\description{
This function allows you be sligthly rude or very rude when greeting.
}
\examples{
hello()
}
\keyword{rude}
