% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rude-package.R
\docType{package}
\name{rude-package}
\alias{rude}
\alias{rude-package}
\title{rude: Rude answering}
\description{
Answers slighly rude or rude to greetings depending in whether the author had a good_day or not.
}
\author{
\strong{Maintainer}: Ynnna AI \email{iliescuaf@gmail.com}

}
\keyword{internal}
